Remove duplicate snackbar code in every ViewModel
