+++ app/src/main/java/com/javainiai/chefskiss/ui/recipescreen/AddRecipeViewModel.kt
@@ -1,18 +1,15 @@
 package com.javainiai.chefskiss.ui.recipescreen
 
 import android.net.Uri
-import androidx.compose.material3.SnackbarDuration
-import androidx.compose.material3.SnackbarHostState
 import androidx.lifecycle.SavedStateHandle
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.javainiai.chefskiss.data.enums.Measurement
 import com.javainiai.chefskiss.data.ingredient.Ingredient
 import com.javainiai.chefskiss.data.recipe.Recipe
 import com.javainiai.chefskiss.data.recipe.RecipesRepository
 import com.javainiai.chefskiss.data.tag.Tag
+import com.javainiai.chefskiss.ui.components.viewmodel.BaseViewModel
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
@@ -47,7 +44,7 @@ data class AddRecipeUiState(
 class AddRecipeViewModel(
     savedStateHandle: SavedStateHandle,
     private val recipesRepository: RecipesRepository
-) : ViewModel() {
+) : BaseViewModel() {
     private val editRecipeId: Long? = savedStateHandle[EditRecipeDestination.editRecipeIdArg]
 
     private var _uiState =
@@ -81,20 +78,6 @@ class AddRecipeViewModel(
     init {
         editRecipeId?.let {
             initializeEditRecipe(it)
-        }
-    }
-
-    val snackbarHostState = SnackbarHostState()
-
-    private var messageInProgress: Job? = null
-    private fun showMessage(message: String) {
-        // cancel in case it hasn't finished so the message can be shown immediately
-        messageInProgress?.cancel()
-        messageInProgress = viewModelScope.launch {
-            snackbarHostState.showSnackbar(
-                message = message,
-                duration = SnackbarDuration.Short
-            )
         }
     }
