+++ app/src/main/java/com/javainiai/chefskiss/ui/shoppinglist/ShoppingListViewModel.kt
@@ -1,15 +1,12 @@
 package com.javainiai.chefskiss.ui.shoppinglist
 
-import androidx.compose.material3.SnackbarDuration
-import androidx.compose.material3.SnackbarHostState
-import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.javainiai.chefskiss.data.recipe.Recipe
 import com.javainiai.chefskiss.data.recipe.RecipeWithIngredients
 import com.javainiai.chefskiss.data.recipe.RecipesRepository
 import com.javainiai.chefskiss.data.recipe.ShopIngredient
 import com.javainiai.chefskiss.data.recipe.ShopRecipe
-import kotlinx.coroutines.Job
+import com.javainiai.chefskiss.ui.components.viewmodel.BaseViewModel
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.first
@@ -22,7 +19,7 @@ data class ShoppingListUiState(
     val recipesWithIngredients: List<RecipeWithIngredients>
 )
 
-class ShoppingListViewModel(private val recipesRepository: RecipesRepository) : ViewModel() {
+class ShoppingListViewModel(private val recipesRepository: RecipesRepository) : BaseViewModel() {
 
     val uiState: StateFlow<ShoppingListUiState> =
         recipesRepository
@@ -47,22 +44,6 @@ class ShoppingListViewModel(private val recipesRepository: RecipesRepository) : 
                 started = SharingStarted.WhileSubscribed(TIMEOUT_MILLIS),
                 initialValue = listOf()
             )
-
-    val snackbarHostState = SnackbarHostState()
-
-    var messageInProgress: Job? = null
-        private set
-
-    private fun showMessage(message: String) {
-        // cancel in case it hasn't finished so the message can be shown immediately
-        messageInProgress?.cancel()
-        messageInProgress = viewModelScope.launch {
-            snackbarHostState.showSnackbar(
-                message = message,
-                duration = SnackbarDuration.Short
-            )
-        }
-    }
 
     fun removeRecipe(recipe: Recipe) {
         viewModelScope.launch {
